requirements.txt
# -------------------------
# Core Framework
# -------------------------
Django>=5.2,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3

# -------------------------
# Config & Security
# -------------------------
python-dotenv>=1.0
django-cors-headers>=4.3

# -------------------------
# Database
# -------------------------
psycopg2-binary>=2.9

# -------------------------
# Async / Queue
# -------------------------
celery>=5.3
celery[redis]>=5.3
redis>=5.0
django-redis>=5.4
